Infer option field type and refactor
